Create a shared LB base class for LBs with multiple children and change ClusgterManagerLoadBalancer to use it.
