+++ xds/src/test/java/io/grpc/xds/ClusterManagerLoadBalancerTest.java
@@ -61,6 +61,7 @@ import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.concurrent.TimeUnit;
 import org.junit.After;
 import org.junit.Before;
@@ -126,10 +127,14 @@ public class ClusterManagerLoadBalancerTest {
     assertThat(pickSubchannel(picker, "childA")).isEqualTo(PickResult.withNoResult());
     assertThat(pickSubchannel(picker, "childB")).isEqualTo(PickResult.withNoResult());
     assertThat(childBalancers).hasSize(2);
-    FakeLoadBalancer childBalancer1 = childBalancers.get(0);
-    FakeLoadBalancer childBalancer2 = childBalancers.get(1);
-    assertThat(childBalancer1.name).isEqualTo("policy_a");
-    assertThat(childBalancer2.name).isEqualTo("policy_b");
+    assertThat(childBalancers.stream()
+        .filter(b -> b.name.equals("policy_a"))
+        .count()).isEqualTo(1);
+    assertThat(childBalancers.stream()
+        .filter(b -> b.name.equals("policy_b"))
+        .count()).isEqualTo(1);
+    FakeLoadBalancer childBalancer1 = getChildBalancerByName("policy_a");
+    FakeLoadBalancer childBalancer2 = getChildBalancerByName("policy_b");
     assertThat(childBalancer1.config).isEqualTo(lbConfigInventory.get("childA"));
     assertThat(childBalancer2.config).isEqualTo(lbConfigInventory.get("childB"));
 
@@ -151,8 +156,7 @@ public class ClusterManagerLoadBalancerTest {
     assertThat(childBalancer2.shutdown).isFalse();
 
     assertThat(childBalancers).hasSize(3);
-    FakeLoadBalancer childBalancer3 = childBalancers.get(2);
-    assertThat(childBalancer3.name).isEqualTo("policy_c");
+    FakeLoadBalancer childBalancer3 = getChildBalancerByName("policy_c");
     assertThat(childBalancer3.config).isEqualTo(lbConfigInventory.get("childC"));
 
     // delayed policy_b deletion
@@ -166,8 +170,8 @@ public class ClusterManagerLoadBalancerTest {
     deliverResolvedAddresses(ImmutableMap.of("childA", "policy_a", "childB", "policy_b"));
 
     assertThat(childBalancers).hasSize(2);
-    FakeLoadBalancer childBalancer1 = childBalancers.get(0);
-    FakeLoadBalancer childBalancer2 = childBalancers.get(1);
+    FakeLoadBalancer childBalancer1 = getChildBalancerByName("policy_a");
+    FakeLoadBalancer childBalancer2 = getChildBalancerByName("policy_b");
     Subchannel subchannel1 = mock(Subchannel.class);
     Subchannel subchannel2 = mock(Subchannel.class);
     childBalancer1.deliverSubchannelState(subchannel1, ConnectivityState.READY);
@@ -184,11 +188,20 @@ public class ClusterManagerLoadBalancerTest {
         .isEqualTo(subchannel2);
   }
 
+  private FakeLoadBalancer getChildBalancerByName(String name) {
+    for (FakeLoadBalancer childLb : childBalancers) {
+      if (childLb.name.equals(name)) {
+        return childLb;
+      }
+    }
+    return null;
+  }
+
   @Test
   public void ignoreBalancingStateUpdateForDeactivatedChildLbs() {
     deliverResolvedAddresses(ImmutableMap.of("childA", "policy_a", "childB", "policy_b"));
     deliverResolvedAddresses(ImmutableMap.of("childB", "policy_b"));
-    FakeLoadBalancer childBalancer1 = childBalancers.get(0);  // policy_a (deactivated)
+    FakeLoadBalancer childBalancer1 = getChildBalancerByName("policy_a");  // policy_a (deactivated)
     Subchannel subchannel = mock(Subchannel.class);
     childBalancer1.deliverSubchannelState(subchannel, ConnectivityState.READY);
     verify(helper, never()).updateBalancingState(
@@ -231,8 +244,8 @@ public class ClusterManagerLoadBalancerTest {
   public void handleNameResolutionError_afterChildLbsInstantiated_propagateToChildLbs() {
     deliverResolvedAddresses(ImmutableMap.of("childA", "policy_a", "childB", "policy_b"));
     assertThat(childBalancers).hasSize(2);
-    FakeLoadBalancer childBalancer1 = childBalancers.get(0);
-    FakeLoadBalancer childBalancer2 = childBalancers.get(1);
+    FakeLoadBalancer childBalancer1 = getChildBalancerByName("policy_a");
+    FakeLoadBalancer childBalancer2 = getChildBalancerByName("policy_b");
     clusterManagerLoadBalancer.handleNameResolutionError(
         Status.UNAVAILABLE.withDescription("resolver error"));
     assertThat(childBalancer1.upstreamError.getCode()).isEqualTo(Code.UNAVAILABLE);
@@ -245,8 +258,8 @@ public class ClusterManagerLoadBalancerTest {
   public void handleNameResolutionError_notPropagateToDeactivatedChildLbs() {
     deliverResolvedAddresses(ImmutableMap.of("childA", "policy_a", "childB", "policy_b"));
     deliverResolvedAddresses(ImmutableMap.of("childB", "policy_b"));
-    FakeLoadBalancer childBalancer1 = childBalancers.get(0);  // policy_a (deactivated)
-    FakeLoadBalancer childBalancer2 = childBalancers.get(1);  // policy_b
+    FakeLoadBalancer childBalancer1 = getChildBalancerByName("policy_a");  // policy_a (deactivated)
+    FakeLoadBalancer childBalancer2 = getChildBalancerByName("policy_b");  // policy_b
     clusterManagerLoadBalancer.handleNameResolutionError(
         Status.UNKNOWN.withDescription("unknown error"));
     assertThat(childBalancer1.upstreamError).isNull();
